pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "levelup"
version = "0.1.0"
description = "AI-Powered Technical Interview Platform"
readme = "README.md"
authors = [
    {name = "LevelUp Team"}
]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.3",
    "httpx>=0.25.1",
    "datasets>=3.1.0",
    "psycopg2-binary>=2.9.10",
    "click",
    "h11",
    "starlette",
    "pydantic_core",
    "typing_extensions",
    "typing_inspection",
    "chromadb>=0.4.15",
    "mistralai>=0.0.12",
    "sentence-transformers>=2.2.2",
    "ruff>=0.1.3",
    "mypy>=1.6.1",
    "black>=23.10.1",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.5.0",
    "numpy>=2.3.0",
    "mlflow==3.1.0",
    "pandas>=2.2.3",
    "scikit-learn>=1.7.0",
    "joblib>=1.5.0",
    "ipykernel>=6.29.5",
    "tqdm>=4.67.1",
    "jupyterlab>=4.4.3",
    "langchain-openai>=0.3.23",
    "langgraph>=0.0.30",
    "langchain>=0.3.25",
    "langgraph>=0.4.8",
    "e2b-code-interpreter",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
]

[tool.setuptools]
packages = ["app", "ml", "data", "pipelines", "tests"]

[tool.ruff]
line-length = 500
target-version = "py311"
exclude = ["data"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app", "ml", "data", "pipelines"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
exclude = ["data/"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
